requirements.txt
python-telegram-bot==21.6
ollamafreeapi
httpx
